feat(servicenow): KSM-1093 modernize integration: FQCN resolver, Yokohama MID libs, CDN download, SBOM publish workflow
